small fixes; checkEventParticipant middleware
